Replace a bunch of stuff
These are breaking changes, but I'm pretty sure nobody uses this, so
it's OK.

This changes how parsing works. The parser now does things
incrementally instead of requiring it to go all at once with a handler
telling it when to stop. The Handler is gone, entirely.

The parser still needs to be rewritten because it's garbage, but that
entire part isn't exported, so it's less likely to be an issue. Just
need to throw some tests in to tell me when behavior changes.

A bufio.Reader is no longer a requirement, though a RuneReader is.  It
helps if the reader supports ReadBytes(delim byte) ([]byte, error),
otherwise this has to read a single byte at a time to determine if it
can read a line. ASCIIReader is implemented as a work-around for the
RuneReader requirement, in case you know you'll only have 7-bit ASCII
in your input.